Delete initialize_heatload
