+++ src/dynamics/thermal/heatload.hpp
@@ -107,4 +107,13 @@ class Heatload {
   void AssertHeatloadParams();
 };
 
+/**
+ * @fn InitHeatload
+ * @brief Initialize Heatload object from csv file
+ * @param[in] time_str: str representing time table, read from csv file
+ * @param[in] internal_heatload_str: str representing internal heatload table, read from csv file
+ * @return Heatload
+ */
+Heatload InitHeatload(const std::vector<std::string>& time_str, const std::vector<std::string>& internal_heatload_str);
+
 #endif  // S2E_DYNAMICS_THERMAL_HEATLOAD_HPP_
